Added Powershell support
